feat(iim-weekly-overview): add --friday, count, and pvariance
This adds a --friday option which lets you specify which friday to base
the report on. This makes it a little easier to run an old report after
bug fixes to see how things changed.

This also adds count and pvariance to the MTT- rows in the Trends
Summary section. This gives more meaning to the mean numbers by adding
an indicator of the divergence from the mean.

+++ src/iim/iim_qbr.py
@@ -52,7 +52,7 @@ def pct_change(a, b):
     return (b - a) / a * 100
 
 
-def tt_in_seconds(tt_value: Optional[timedelta]) -> Optional[int]:
+def tt_in_seconds(tt_value: Optional[timedelta]) -> Optional[float]:
     if tt_value is None:
         return tt_value
